tidy up initIdeProjectLogo
